+++ business/src/main/java/com/emotionbank/business/api/account/dto/request/AccountRequestDto.java
@@ -1,0 +1,4 @@
+package com.emotionbank.business.api.account.dto.request;
+
+public class AccountRequestDto {
+}
