+++ business/src/main/java/com/emotionbank/business/api/account/dto/response/AccountResponseDto.java
@@ -1,0 +1,4 @@
+package com.emotionbank.business.api.account.dto.response;
+
+public class AccountResponseDto {
+}
