+++ business/src/main/java/com/emotionbank/business/api/transaction/dto/request/TransactionRequestDto.java
@@ -1,0 +1,4 @@
+package com.emotionbank.business.api.transaction.dto.request;
+
+public class TransactionRequestDto {
+}
