+++ business/src/main/java/com/emotionbank/business/api/transaction/dto/response/TransactionResponseDto.java
@@ -1,0 +1,4 @@
+package com.emotionbank.business.api.transaction.dto.response;
+
+public class TransactionResponseDto {
+}
