+++ business/src/main/java/com/emotionbank/business/api/user/dto/request/UserRequestDto.java
@@ -1,0 +1,4 @@
+package com.emotionbank.business.api.user.dto.request;
+
+public class UserRequestDto {
+}
