+++ business/src/main/java/com/emotionbank/business/api/user/dto/response/UserResponseDto.java
@@ -1,0 +1,4 @@
+package com.emotionbank.business.api.user.dto.response;
+
+public class UserResponseDto {
+}
