+++ business/src/main/java/com/emotionbank/business/domain/account/dto/AccountDto.java
@@ -1,0 +1,5 @@
+package com.emotionbank.business.domain.account.dto;
+
+public class AccountDto
+{
+}
