--- business/src/main/java/com/emotionbank/business/domain/account/dto/request/AccountRequestDto.java
@@ -1,4 +1,0 @@
-package com.emotionbank.business.domain.account.dto.request;
-
-public class AccountRequestDto {
-}
