--- business/src/main/java/com/emotionbank/business/domain/account/dto/response/AccountResponseDto.java
@@ -1,4 +1,0 @@
-package com.emotionbank.business.domain.account.dto.response;
-
-public class AccountResponseDto {
-}
