+++ business/src/main/java/com/emotionbank/business/domain/transaction/dto/TransactionDto.java
@@ -1,0 +1,4 @@
+package com.emotionbank.business.domain.transaction.dto;
+
+public class TransactionDto {
+}
