--- business/src/main/java/com/emotionbank/business/domain/transaction/dto/request/TransactionRequestDto.java
@@ -1,4 +1,0 @@
-package com.emotionbank.business.domain.transaction.dto.request;
-
-public class TransactionRequestDto {
-}
