--- business/src/main/java/com/emotionbank/business/domain/transaction/dto/response/TransactionResponseDto.java
@@ -1,4 +1,0 @@
-package com.emotionbank.business.domain.transaction.dto.response;
-
-public class TransactionResponseDto {
-}
