+++ business/src/main/java/com/emotionbank/business/domain/user/dto/UserDto.java
@@ -1,0 +1,4 @@
+package com.emotionbank.business.domain.user.dto;
+
+public class UserDto {
+}
