--- business/src/main/java/com/emotionbank/business/domain/user/dto/request/UserRequestDto.java
@@ -1,4 +1,0 @@
-package com.emotionbank.business.domain.user.dto.request;
-
-public class UserRequestDto {
-}
