--- business/src/main/java/com/emotionbank/business/domain/user/dto/response/UserResponseDto.java
@@ -1,4 +1,0 @@
-package com.emotionbank.business.domain.user.dto.response;
-
-public class UserResponseDto {
-}
